[PR-17] Remmend API 연동 - Path를 리스트로 출력 (#17)
* [PR-17] UserPath를 보여줄 수 있도록 API 수정 및 ViewModel 함수 구현

* [PR-17] User 생성, 위치 업데이트, 경로 추천 받아오기 API 연동

* [PR-17] Add EOF to UserInsertResponse, Point

+++ android/app/src/main/java/kr/pnu/ga2019/data/network/request/UpdateUserLocationRequest.kt
@@ -12,8 +12,8 @@ data class UpdateUserLocationRequest(
     val memberSeq: Int,
 
     @SerializedName("locationX")
-    val locationX: Double,
+    val locationX: Int,
 
     @SerializedName("locationY")
-    val locationY: Double
+    val locationY: Int
 ) : Request
